got rid of error in file deletion

+++ src/RandomClusteringGraph.jl
@@ -52,7 +52,7 @@ function random_clustering_graph(degs::Array{Int,1},N::Int,C::AbstractFloat,dele
     run(`java -jar RandomClusteringNetwork.jar $deg_seq_filename $N $C $out_filename`)
     run(`rm $(deg_seq_filename)`) 
     G =  read_edgelist(out_filename)
-    if delete_out
+    if delete_out && isfile(out_filename)
         run(`rm $(out_filename)`) 
     end
     return G
